requirements.txt
Django==2.1
django-appconf==1.0.2
django-basicauth==0.5.1
django-compressor==2.2
django-sass-processor==0.7
djangorestframework==3.8.2
libsass==0.14.5
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
virtualenv==16.0.0
